backend/requirements.txt
asgiref==3.10.0
attrs==25.4.0
Django==5.2.7
django-ckeditor-5==0.2.18
django-cors-headers==4.9.0
django-environ==0.12.0
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.28.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
